Add findMissing and reportMissing routes, update navbar, and enhance user model, enhance Dashboard
