+++ Frontend/app/report/page.tsx
@@ -1,157 +1,183 @@
 "use client";
 
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import { Textarea } from "@/components/ui/textarea";
-import { Upload, Camera, X } from "lucide-react";
+import { Upload, SearchIcon, X } from "lucide-react";
 import Image from "next/image";
 
-export default function ReportPage() {
-  const [previewUrls, setPreviewUrls] = useState<string[]>([]);
-  const [selectedFiles, setSelectedFiles] = useState<File[]>([]);
+// Define types
+interface SearchResult {
+  _id: string;
+  name: string;
+  age: number;
+  missingDate: string;
+  lastSeenLocation: string;
+  photos: string[];
+}
 
-  const [formData, setFormData] = useState({
-    firstName: "",
-    lastName: "",
-    age: "",
-    gender: "",
-    lastSeenLocation: "",
-    missingDate: "",
+interface SearchParams {
+  name: string;
+  description: string;
+  location: string;
+}
+
+export default function ReportPage() {
+  const [searchResults, setSearchResults] = useState<SearchResult[]>([]);
+  const [searchParams, setSearchParams] = useState<SearchParams>({
+    name: "",
     description: "",
+    location: "",
   });
 
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
-    setFormData({ ...formData, [e.target.id]: e.target.value });
-  };
+  const [selectedFiles, setSelectedFiles] = useState<File[]>([]);
+  const [previewUrls, setPreviewUrls] = useState<string[]>([]);
+  const [accessToken, setAccessToken] = useState<string | null>(null);
 
-  const handleImageUpload = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const files = e.target.files;
-    if (files) {
-      const newFiles = Array.from(files);
-      setSelectedFiles([...selectedFiles, ...newFiles]);
-      const newUrls = newFiles.map((file) => URL.createObjectURL(file));
-      setPreviewUrls([...previewUrls, ...newUrls]);
+  // Fetch access token from localStorage
+  useEffect(() => {
+    const token = localStorage.getItem("accessToken");
+    if (token) {
+      setAccessToken(token);
     }
+  }, []);
+
+  // Handle input change
+  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const { name, value } = e.target;
+    setSearchParams((prev) => ({ ...prev, [name]: value }));
   };
 
-  const handleRemoveImage = (index: number) => {
-    setPreviewUrls(previewUrls.filter((_, i) => i !== index));
-    setSelectedFiles(selectedFiles.filter((_, i) => i !== index));
+  // Handle multiple file selection
+  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+  const files = Array.from(e.target.files || []);
+  if (files.length) {
+    setSelectedFiles((prevFiles) => [...prevFiles, ...files]);
+    setPreviewUrls((prevUrls) => [...prevUrls, ...files.map((file) => URL.createObjectURL(file))]);
+  }
+};
+
+
+  // Remove a selected image
+  const removeImage = (index: number) => {
+    setSelectedFiles((prevFiles) => prevFiles.filter((_, i) => i !== index));
+    setPreviewUrls((prevUrls) => prevUrls.filter((_, i) => i !== index));
   };
 
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-
-    const accessToken = localStorage.getItem("accessToken");
-
-    if (!accessToken) {
-      alert("You must be logged in to submit a report.");
-      return;
-    }
-
-    const reportData = new FormData();
-    reportData.append("name", `${formData.firstName} ${formData.lastName}`);
-    reportData.append("age", formData.age);
-    reportData.append("gender", formData.gender);
-    reportData.append("missingDate", formData.missingDate);
-    reportData.append("lastSeenLocation", formData.lastSeenLocation);
-    reportData.append("description", formData.description);
-
-    selectedFiles.forEach((file) => {
-      reportData.append("photos", file);
-    });
-
+  // Handle search request (POST request to create sighting report)
+  const handleSearch = async () => {
     try {
-      const response = await fetch("http://localhost:5000/api/v1/missing-persons/", {
-        method: "POST",
-        body: reportData,
-        headers: {
-          Authorization: `Bearer ${accessToken}`,
-        },
+      const formData = new FormData();
+
+      // Append text fields if they have values
+      Object.entries(searchParams).forEach(([key, value]) => {
+        if (value) formData.append(key, value);
       });
 
-      console.log("Raw Response:", response);
-      const data = await response.json();
+      // Append multiple images
+      selectedFiles.forEach((file) => {
+        formData.append("photos", file);
+      });
 
-      if (response.ok) {
-        alert("Missing person report submitted successfully!");
-        setFormData({
-          firstName: "",
-          lastName: "",
-          age: "",
-          gender: "",
-          lastSeenLocation: "",
-          missingDate: "",
-          description: "",
-        });
-        setPreviewUrls([]);
-        setSelectedFiles([]);
-      } else {
-        alert(data || "Failed to submit report.");
+      const response = await fetch("http://localhost:5000/api/v1/sightings", {
+        method: "POST",
+        headers: {
+          Authorization: accessToken ? `Bearer ${accessToken}` : "",
+        },
+        body: formData, // FormData automatically sets the correct headers
+      });
+
+      if (!response.ok) {
+        const errorMessage = await response.text();
+        throw new Error(errorMessage || "Failed to submit sighting report");
       }
+
+      const data = await response.json();
+      setSearchResults((prev) => [...prev, data.data as SearchResult]);
+      setSearchParams({ name: "", description: "", location: "" });  
+      setSelectedFiles([]);
+      setPreviewUrls([]);
     } catch (error) {
-      console.error("Error submitting report:", error);
-      alert("An error occurred. Please try again.");
+      console.error("Search error:", error);
+      alert("Error: " + error);
     }
   };
 
   return (
-    <div className="container max-w-7xl px-4 sm:px-6 lg:px-40 py-10">
-      <form onSubmit={handleSubmit} className="space-y-8">
-        <div className="text-center">
-          <h1 className="text-3xl font-bold tracking-tight">Report Missing Person</h1>
-          <p className="text-muted-foreground mt-2">Please provide as much information as possible.</p>
-        </div>
+    <div className="container py-10">
+      <div className="max-w-4xl mx-auto space-y-6">
+        <h1 className="text-3xl font-bold tracking-tight text-center">Report Missing Person</h1>
+        <p className="text-muted-foreground mt-2 text-center">
+          Provide details to report a missing person sighting.
+        </p>
 
-        {/* Personal Information */}
-        <div className="space-y-6">
-          <h2 className="text-2xl font-semibold">Personal Information</h2>
-          <div className="grid gap-6 sm:grid-cols-2">
-            <Input id="firstName" placeholder="First Name" onChange={handleChange} />
-            <Input id="lastName" placeholder="Last Name" onChange={handleChange} />
-            <Input id="age" type="number" placeholder="Age" onChange={handleChange} />
-            <Input id="gender" placeholder="Gender" onChange={handleChange} />
+        {/* Input Fields */}
+        <div className="grid gap-4 sm:grid-cols-2">
+          <div className="space-y-2">
+            <Label htmlFor="name">Name</Label>
+            <Input id="name" name="name" placeholder="Enter name (optional)" value={searchParams.name} onChange={handleInputChange} />
+          </div>
+          <div className="space-y-2">
+            <Label htmlFor="location">Location</Label>
+            <Input id="location" name="location" placeholder="Enter sighting location" value={searchParams.location} onChange={handleInputChange} />
+          </div>
+          <div className="space-y-2 sm:col-span-2">
+            <Label htmlFor="description">Description</Label>
+            <Input id="description" name="description" placeholder="Additional details" value={searchParams.description} onChange={handleInputChange} />
           </div>
         </div>
 
-        {/* Last Seen Information */}
-        <div className="space-y-6">
-          <h2 className="text-2xl font-semibold">Last Seen Information</h2>
-          <Input id="lastSeenLocation" placeholder="Last Known Location" onChange={handleChange} />
-          <Input id="missingDate" type="date" onChange={handleChange} />
-          <Textarea id="description" placeholder="Description..." rows={4} onChange={handleChange} />
-        </div>
-
-        {/* Photo Upload */}
-        <div className="space-y-6">
-          <h2 className="text-2xl font-semibold">Photos</h2>
-          <div className="flex flex-wrap gap-4">
-            {previewUrls.map((url, index) => (
-              <div key={index} className="relative w-32 h-32">
-                <Image src={url} alt="Preview" layout="fill" objectFit="cover" className="rounded-lg" />
-                <button
-                  type="button"
-                  className="absolute top-0 right-0 bg-red-500 text-white rounded-full p-1"
-                  onClick={() => handleRemoveImage(index)}
-                >
-                  <X className="h-5 w-5" />
-                </button>
+        {/* Upload Section */}
+        <div className="mt-6">
+          <label
+            htmlFor="searchPhoto"
+            className="flex flex-col items-center justify-center w-full h-64 border-2 border-dashed rounded-lg cursor-pointer hover:bg-muted/50"
+          >
+            {previewUrls.length > 0 ? (
+              <div className="grid grid-cols-3 gap-2 p-2">
+                {previewUrls.map((url, index) => (
+                  <div key={index} className="relative">
+                    <Image
+                      src={url}
+                      alt={`Selected file ${index + 1}`}
+                      className="h-32 w-32 object-cover rounded-lg"
+                      height={32}
+                      width={32}
+                    />
+                    <button
+                      type="button"
+                      onClick={() => removeImage(index)}
+                      className="absolute top-1 right-1 bg-red-500 text-white p-1 rounded-full hover:bg-red-600"
+                    >
+                      <X className="h-4 w-4" />
+                    </button>
+                  </div>
+                ))}
               </div>
-            ))}
-          </div>
-          <label className="flex items-center gap-4 cursor-pointer border-2 border-dashed p-4 rounded-lg hover:bg-muted/50">
-            <Upload className="h-8 w-8 text-muted-foreground" />
-            <span>Click to upload photos</span>
-            <input id="photo" type="file" className="hidden" accept="image/*" multiple onChange={handleImageUpload} />
+            ) : (
+              <div className="flex flex-col items-center justify-center pt-5 pb-6">
+                <Upload className="h-12 w-12 mb-3 text-muted-foreground" />
+                <p className="mb-2 text-sm text-muted-foreground">
+                  <span className="font-semibold">Click to upload</span> or drag and drop
+                </p>
+                <p className="text-xs text-muted-foreground">
+                  Supported formats: JPG, PNG (MAX. 10MB)
+                </p>
+              </div>
+            )}
+            <input id="searchPhoto" type="file" className="hidden" accept="image/*" multiple onChange={handleFileChange} />
           </label>
         </div>
 
-        <Button type="submit" size="lg" className="w-full">
-          Submit Report
-        </Button>
-      </form>
+        {/* Submit Button */}
+        <div className="flex justify-center">
+          <Button size="lg" className="gap-2" onClick={handleSearch}>
+            <SearchIcon className="h-4 w-4" />
+            Submit
+          </Button>
+        </div>
+      </div>
     </div>
   );
 }
